feat: added a verification to know if the barbershp id exists.

+++ src/modules/barbershops/services/delete-barbershop.service.ts
@@ -1,5 +1,8 @@
 import { barbershopRepository } from "../repositories/barbershop.repository";
 
 export const deleteBarbershopService = async (id: string) => {
+  if (!id) {
+    throw new Error("ID is required");
+  }
   return barbershopRepository.delete(id);
 };
